fix decision calc

+++ database/migrations/2025_06_08_124241_create_criteria_weights_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('criteria_weights', function (Blueprint $table) {
             $table->id();
             $table->string('criteria');
-            $table->integer('weight');
+            $table->float('weight');
             $table->boolean('is_active')->default(true);
             $table->timestamps();
         });
